DQL/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
cloudpickle==3.0.0
colorama==0.4.6
filelock==3.15.4
fsspec==2024.6.1
gym==0.26.2
gym-notices==0.0.8
gym-super-mario-bros==7.4.0
idna==3.7
importlib_metadata==8.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
nes_py==8.2.1
networkx==3.1
numpy==1.24.4
packaging==24.1
pillow==10.4.0
pyglet==1.5.21
requests==2.32.3
sympy==1.13.1
tensordict==0.3.0
torch==2.2.0
torchaudio==2.2.0
torchrl==0.3.0
torchvision==0.17.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
zipp==3.19.2
opencv-python==4.10.0.84
tensorboard==2.14.0
h5py==3.11.0
argparse==1.4.0